+++ docs/fr_FR/changelog.md
@@ -2,6 +2,7 @@
 
 ## 4.2.0
 
+- **Synthèse** : Possibilité de paramétrage des objets pour aller vers un *design* ou une *vue* depuis la synthèse.
 - **Logs** : Coloration des logs en fonction de certains évènements. Possibilité d'afficher le log brut.
 - **Analyse / Equipements** : Les commandes orphelines affichent maintenant leurs noms et date de suppression si encore dans l'historique de suppression.
 - **Plugins / Gestion** : Affichage de la catégorie du plugin, et d'un lien pour ouvrir directement la page de celui-ci.